Renames what didn't end up being a whole SerialPort.
