+++ Machines/AtariST/IntelligentKeyboard.hpp
@@ -10,7 +10,7 @@
 #define IntelligentKeyboard_hpp
 
 #include "../../ClockReceiver/ClockingHintSource.hpp"
-#include "../../Components/SerialPort/SerialPort.hpp"
+#include "../../Components/Serial/Line.hpp"
 #include "../KeyboardMachine.hpp"
 
 #include "../../Inputs/Mouse.hpp"
